+++ Graphics/vtkLineSource.h
@@ -32,13 +32,13 @@ public:
 
   // Description:
   // Set position of first end point.
-  vtkSetVector3Macro(Point1,float);
-  vtkGetVectorMacro(Point1,float,3);
+  vtkSetVector3Macro(Point1,double);
+  vtkGetVectorMacro(Point1,double,3);
 
   // Description:
   // Set position of other end point.
-  vtkSetVector3Macro(Point2,float);
-  vtkGetVectorMacro(Point2,float,3);
+  vtkSetVector3Macro(Point2,double);
+  vtkGetVectorMacro(Point2,double,3);
 
   // Description:
   // Divide line into resolution number of pieces.
@@ -50,8 +50,8 @@ protected:
   ~vtkLineSource() {};
 
   void Execute();
-  float Point1[3];
-  float Point2[3];
+  double Point1[3];
+  double Point2[3];
   int Resolution;
 private:
   vtkLineSource(const vtkLineSource&);  // Not implemented.
